+++ api/src/main/java/tennisclub/dto/court/CourtDto.java
@@ -1,4 +1,4 @@
-package tennisclub.dto;
+package tennisclub.dto.court;
 
 import tennisclub.enums.CourtType;
