+++ api/src/main/java/tennisclub/dto/lesson/LessonWithCourtDTO.java
@@ -1,37 +1,24 @@
-package tennisclub.dto;
+package tennisclub.dto.lesson;
 
-import javax.validation.constraints.*;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-public class TournamentCreateDTO {
-    @NotNull
-    private CourtDto court;
-
-    @NotNull
-    @FutureOrPresent
+public class LessonWithCourtDTO {
+    private Long id;
+    // TODO: uncomment once the PR is merged :)
+    // private CourtDto court;
     private LocalDateTime startTime;
-
-    @NotNull
-    @Future
     private LocalDateTime endTime;
-
-    @NotBlank
-    private String name;
-
-    @Min(0)
     private Integer capacity;
+    // TODO: uncomment once the PR is merged :)
+    // private Level level;
 
-    @NotNull
-    @Min(0)
-    private Integer prize;
-
-    public CourtDto getCourt() {
-        return court;
+    public Long getId() {
+        return id;
     }
 
-    public void setCourt(CourtDto court) {
-        this.court = court;
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public LocalDateTime getStartTime() {
@@ -50,14 +37,6 @@ public class TournamentCreateDTO {
         this.endTime = endTime;
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
     public Integer getCapacity() {
         return capacity;
     }
@@ -66,29 +45,21 @@ public class TournamentCreateDTO {
         this.capacity = capacity;
     }
 
-    public Integer getPrize() {
-        return prize;
-    }
-
-    public void setPrize(Integer prize) {
-        this.prize = prize;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        TournamentCreateDTO tournamentDTO = (TournamentCreateDTO) o;
-        return Objects.equals(court, tournamentDTO.getCourt()) &&
-                Objects.equals(startTime, tournamentDTO.getStartTime())
-                && Objects.equals(endTime, tournamentDTO.getEndTime());
+        LessonWithCourtDTO lessonDTO = (LessonWithCourtDTO) o;
+        return // Objects.equals(court, lessonDTO.getCourt());
+                Objects.equals(startTime, lessonDTO.getStartTime())
+                && Objects.equals(endTime, lessonDTO.getEndTime());
     }
 
     @Override
     public int hashCode() {
         final int prime = 31;
         int result = 1;
-        result = prime * result  + ((court == null) ? 0 : court.hashCode());
+        // result = prime * result  + ((court == null) ? 0 : court.hashCode());
         result = prime * result  + ((startTime == null) ? 0 : startTime.hashCode());
         result = prime * result  + ((endTime == null) ? 0 : endTime.hashCode());
         return result;
