+++ service/src/main/java/tennisclub/facade/CourtFacadeImpl.java
@@ -3,11 +3,11 @@ package tennisclub.facade;
 import org.dozer.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import tennisclub.dto.CourtDto;
+import tennisclub.dto.court.CourtDto;
 import tennisclub.entity.Court;
 import tennisclub.enums.CourtType;
 import tennisclub.service.CourtService;
-import tennisclub.dto.CourtCreateDto;
+import tennisclub.dto.court.CourtCreateDto;
 
 import java.util.List;
 import java.util.stream.Collectors;
